more cleanup - taking out the boiler plate from the service into the middleware directly
